requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
playwright==1.43.0
pyee==11.1.0
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
typing_extensions==4.11.0
urllib3==2.2.1
